Moved browseSearches to its own file
